Develop to main. Release v2

+++ OpenEdX/AppDelegate.swift
@@ -7,15 +7,23 @@
 
 import UIKit
 import Core
+import OEXFoundation
 import Swinject
 import Profile
 import GoogleSignIn
 import FacebookCore
 import MSAL
+import UserNotifications
+import OEXFirebaseAnalytics
+import FirebaseCore
+import FirebaseMessaging
 import Theme
+import BackgroundTasks
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
+    
+    static let bgAppTaskId = "openEdx.offlineProgressSync"
     
     static var shared: AppDelegate {
         UIApplication.shared.delegate as! AppDelegate
@@ -23,6 +31,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     var window: UIWindow?
         
+    private let pluginManager = PluginManager()
     private var assembler: Assembler?
     
     private var lastForceLogoutTime: TimeInterval = 0
@@ -32,6 +41,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
     ) -> Bool {
         initDI()
+        initPlugins()
+        
         if let config = Container.shared.resolve(ConfigProtocol.self) {
             Theme.Shapes.isRoundedCorners = config.theme.isRoundedCorners
             
@@ -42,6 +53,20 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                 )
             }
             configureDeepLinkServices(launchOptions: launchOptions)
+            
+            let pushManager = Container.shared.resolve(PushNotificationsManager.self)
+            
+            if config.firebase.enabled {
+                FirebaseApp.configure()
+                if config.firebase.cloudMessagingEnabled {
+                    Messaging.messaging().delegate = pushManager
+                    UNUserNotificationCenter.current().delegate = pushManager
+                }
+            }
+            
+            if pushManager?.hasProviders == true {
+                UIApplication.shared.registerForRemoteNotifications()
+            }
         }
 
         Theme.Fonts.registerFonts()
@@ -49,17 +74,20 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         window?.rootViewController = RouteController()
         window?.makeKeyAndVisible()
         window?.tintColor = Theme.UIColors.accentColor
-        
+          
         NotificationCenter.default.addObserver(
             self,
-            selector: #selector(forceLogoutUser),
-            name: .onTokenRefreshFailed,
+            selector: #selector(didUserAuthorize),
+            name: .userAuthorized,
             object: nil
         )
         
-        if let pushManager = Container.shared.resolve(PushNotificationsManager.self) {
-            pushManager.performRegistration()
-        }
+        NotificationCenter.default.addObserver(
+            self,
+            selector: #selector(didUserLogout),
+            name: .userLoggedOut,
+            object: nil
+        )
 
         return true
     }
@@ -105,6 +133,15 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
         return false
     }
+    
+    private func initPlugins() {
+        guard let config = Container.shared.resolve(ConfigProtocol.self) else { return }
+        if config.firebase.enabled && config.firebase.isAnalyticsSourceFirebase {
+            pluginManager.addPlugin(analyticsService: FirebaseAnalyticsService())
+        }
+        
+        // Initialize your plugins here
+    }
 
     private func initDI() {
         let navigation = UINavigationController()
@@ -112,7 +149,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         
         assembler = Assembler(
             [
-                AppAssembly(navigation: navigation),
+                AppAssembly(navigation: navigation, pluginManager: pluginManager),
                 NetworkAssembly(),
                 ScreenAssembly()
             ],
@@ -120,21 +157,32 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         )
     }
     
-    @objc private func forceLogoutUser() {
+    @objc private func didUserAuthorize() {
+        Container.shared.resolve(PushNotificationsManager.self)?.synchronizeToken()
+    }
+    
+    @objc func didUserLogout(_ notification: Notification) {
         guard Date().timeIntervalSince1970 - lastForceLogoutTime > 5 else {
             return
         }
-        let analyticsManager = Container.shared.resolve(AnalyticsManager.self)
-        analyticsManager?.userLogout(force: true)
-        
-        lastForceLogoutTime = Date().timeIntervalSince1970
-        
-        Container.shared.resolve(CoreStorage.self)?.clear()
-        Task {
-            await Container.shared.resolve(DownloadManagerProtocol.self)?.deleteAllFiles()
+        if let userInfo = notification.userInfo,
+           userInfo[Notification.UserInfoKey.isForced] as? Bool == true {
+            let analyticsManager = Container.shared.resolve(AnalyticsManager.self)
+            analyticsManager?.userLogout(force: true)
+            
+            lastForceLogoutTime = Date().timeIntervalSince1970
+            
+            Container.shared.resolve(CoreStorage.self)?.clear()
+            Container.shared.resolve(CorePersistenceProtocol.self)?.deleteAllProgress()
+            Task {
+                await Container.shared.resolve(DownloadManagerProtocol.self)?.deleteAllFiles()
+            }
+            Container.shared.resolve(CoreDataHandlerProtocol.self)?.clear()
+            window?.rootViewController = RouteController()
         }
-        Container.shared.resolve(CoreDataHandlerProtocol.self)?.clear()
-        window?.rootViewController = RouteController()
+        
+        UNUserNotificationCenter.current().removeAllDeliveredNotifications()
+        Container.shared.resolve(PushNotificationsManager.self)?.refreshToken()
     }
     
     // Push Notifications
@@ -151,8 +199,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         didReceiveRemoteNotification userInfo: [AnyHashable: Any],
         fetchCompletionHandler completionHandler: @escaping (UIBackgroundFetchResult) -> Void
     ) {
-        guard let pushManager = Container.shared.resolve(PushNotificationsManager.self)
-        else {
+        guard let pushManager = Container.shared.resolve(PushNotificationsManager.self) else {
             completionHandler(.newData)
             return
         }
@@ -164,5 +211,47 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func configureDeepLinkServices(launchOptions: [UIApplication.LaunchOptionsKey: Any]?) {
         guard let deepLinkManager = Container.shared.resolve(DeepLinkManager.self) else { return }
         deepLinkManager.configureDeepLinkService(launchOptions: launchOptions)
+    }
+    
+    // Background progress update
+    
+    func registerBackgroundTask() {
+        let isRegistered = BGTaskScheduler.shared.register(
+            forTaskWithIdentifier: Self.bgAppTaskId,
+            using: nil
+        ) { task in
+            debugLog("Background task is executing: \(task.identifier)")
+            guard let task = task as? BGAppRefreshTask else { return }
+            self.handleAppRefreshTask(task: task)
+        }
+        debugLog("Is the background task registered? \(isRegistered)")
+    }
+    
+    func handleAppRefreshTask(task: BGAppRefreshTask) {
+        //In real case scenario we should check internet here
+        reScheduleAppRefresh()
+        
+        task.expirationHandler = {
+            //This Block call by System
+            //Canel your all tak's & queues
+            task.setTaskCompleted(success: true)
+        }
+        
+        let offlineSyncManager = Container.shared.resolve(OfflineSyncManagerProtocol.self)!
+        Task {
+            await offlineSyncManager.syncOfflineProgress()
+            task.setTaskCompleted(success: true)
+        }
+    }
+    
+    func reScheduleAppRefresh() {
+        let request = BGAppRefreshTaskRequest(identifier: Self.bgAppTaskId)
+        request.earliestBeginDate = Date(timeIntervalSinceNow: 60 * 60) // App Refresh after 60 minute.
+        //Note :: EarliestBeginDate should not be set to too far into the future.
+        do {
+            try BGTaskScheduler.shared.submit(request)
+        } catch {
+            debugLog("Could not schedule app refresh: \(error)")
+        }
     }
 }
